# backend/swagger.yaml
openapi: 3.0.3
info:
  title: TaskFlow API
  description: |
    Full-stack Project & Task Management Platform API.
    
    ## Authentication
    Most endpoints require a JWT access token in the `Authorization: Bearer <token>` header.
    
    ## Roles
    - **ADMIN** — Full access to all resources
    - **PROJECT_MANAGER** — Manage own projects and their tasks
    - **TEAM_MEMBER** — Manage assigned tasks only
  version: 1.0.0
  contact:
    name: TaskFlow Team
servers:
  - url: http://localhost:5000/api
    description: Development server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token (15min expiry)
...
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        default: 1
        minimum: 1
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
paths:
...
  /projects:
    get:
      tags: [Projects]
      summary: List projects
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: status
          in: query
          schema:
            type: string
            enum: [ACTIVE, COMPLETED, ON_HOLD]
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Paginated projects
    post:
      tags: [Projects]
      summary: Create project
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, startDate]
              properties:
                name:
                  type: string
                  minLength: 2
                  maxLength: 200
                description:
                  type: string
                  maxLength: 2000
                startDate:
                  type: string
                  format: date
                endDate:
                  type: string
                  format: date
                status:
                  type: string
                  enum: [ACTIVE, COMPLETED, ON_HOLD]
      responses:
        '201':
          description: Project created
  /projects/{id}:
    get:
      tags: [Projects]
      summary: Get project details
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Project details
        '403':
          description: Not a member
        '404':
          description: Project not found
    put:
      tags: [Projects]
      summary: Update project
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                status:
                  type: string
                  enum: [ACTIVE, COMPLETED, ON_HOLD]
                startDate:
                  type: string
                  format: date
                endDate:
                  type: string
                  format: date
      responses:
        '200':
          description: Project updated
    delete:
      tags: [Projects]
      summary: Delete project
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Project deleted
  /projects/{id}/members:
    post:
      tags: [Projects]
      summary: Add member to project
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userId]
              properties:
                userId:
                  type: string
                role:
                  type: string
                  default: member
      responses:
        '201':
          description: Member added
        '409':
          description: User is already a member
  /projects/{id}/members/{userId}:
    delete:
      tags: [Projects]
      summary: Remove member from project
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Member removed
  /projects/{projectId}/tasks:
    get:
      tags: [Tasks]
      summary: List project tasks
      security:
        - bearerAuth: []
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: status
          in: query
          schema:
            type: string
            enum: [TODO, IN_PROGRESS, IN_REVIEW, DONE]
        - name: priority
          in: query
          schema:
            type: string
            enum: [LOW, MEDIUM, HIGH, URGENT]
        - name: assigneeId
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Paginated tasks
    post:
      tags: [Tasks]
      summary: Create task
      security:
        - bearerAuth: []
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title]
              properties:
                title:
                  type: string
                  minLength: 2
                  maxLength: 200
                description:
                  type: string
                  maxLength: 5000
                priority:
                  type: string
                  enum: [LOW, MEDIUM, HIGH, URGENT]
                  default: MEDIUM
                dueDate:
                  type: string
                  format: date-time
                assigneeId:
                  type: string
                labelIds:
                  type: array
                  items:
                    type: string
      responses:
        '201':
          description: Task created
  /tasks/{id}:
    get:
      tags: [Tasks]
      summary: Get task details
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Task details with subtasks, labels, comments
    put:
      tags: [Tasks]
      summary: Update task
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                description:
                  type: string
                priority:
                  type: string
                  enum: [LOW, MEDIUM, HIGH, URGENT]
                dueDate:
                  type: string
                  format: date-time
                assigneeId:
                  type: string
                labelIds:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: Task updated
    delete:
      tags: [Tasks]
      summary: Delete task
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Task deleted
  /tasks/{id}/status:
    patch:
      tags: [Tasks]
      summary: Update task status
      security:
        - bearerAuth: []
      description: |
        Linear workflow: TODO → IN_PROGRESS → IN_REVIEW → DONE
        
        - ADMIN can skip stages
        - Others must follow linear progression
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [TODO, IN_PROGRESS, IN_REVIEW, DONE]
      responses:
        '200':
          description: Status updated
        '400':
          description: Cannot skip stages
  /tasks/{id}/subtasks:
    post:
      tags: [Tasks]
      summary: Add subtask
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title]
              properties:
                title:
                  type: string
                  minLength: 1
                  maxLength: 200
      responses:
        '201':
          description: Subtask created
  /tasks/{id}/subtasks/{subtaskId}:
    patch:
      tags: [Tasks]
      summary: Toggle subtask completion
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: subtaskId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [isCompleted]
              properties:
                isCompleted:
                  type: boolean
      responses:
        '200':
          description: Subtask updated
    delete:
      tags: [Tasks]
      summary: Delete subtask
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: subtaskId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Subtask deleted
...